requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.20
langsmith>=0.0.77

# Vector Store
qdrant-client>=1.7.0

# Retrieval & Reranking
rank-bm25>=0.2.2
cohere>=4.37

# External APIs
tavily-python>=0.3.0
openai>=1.6.1

# Evaluation
ragas>=0.1.0
datasets>=2.15.0

# Web Framework
streamlit>=1.29.0
streamlit-folium>=0.15.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
folium>=0.15.0
pydeck>=0.8.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Utilities
pyyaml>=6.0.0
requests>=2.31.0
aiohttp>=3.9.0
python-dateutil>=2.8.0

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

flask>=3.1.2
ollama>=0.6.1
asyncpg>=0.31.0
dotenv>=1.2.1
